Move back AvmString into ruffle_core::string
This allows removing the gc_arena dependency in ruffle_wstr
